requirements.txt

fastapi>=0.100.0
openai>=1.0.0
pydantic>=2.0.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
